requirements.txt
discord.py==1.4.1
requests==2.23.0
gspread==3.6.0
oauth2client==4.1.3